update pk type enoe geo

+++ etl/enoe/dim_geo_inegi_enoe_v2.py
@@ -32,7 +32,7 @@ class DimEnoeGeoV2Pipeline(EasyPipeline):
             'mun_slug':                 'String',
             'cve_mun':                  'String',
             'cve_mun_full':             'String', 
-            'code':                     'UInt32'
+            'code':                     'UInt64'
         }
 
         download_step = DownloadStep(
